backend/requirements.txt
fastapi[standard]>=0.115.0
sqlmodel>=0.0.22
python-dotenv>=1.0.1
psycopg2-binary>=2.9.10
pydantic-settings>=2.6.0
groq>=0.11.0
